Create README.md
Elementix is the ultimate app for exploring and mastering the periodic table of elements. Packed with comprehensive information on every element, interactive visualizations, and built-in quizzes, the app makes learning about the elements fun, interactive, and engaging. With search and filter functionality, customizable themes, and user-friendly navigation, you can easily access detailed descriptions and studies of the different elements on the periodic table. Designed for students, educators, and researchers alike, Elementix offers a comprehensive guide to understanding the world of chemistry. Discover a new way to explore and learn about the elements with Elementix.